Removing MOS Files.
